Remove shortening and fix mimimum light level
I was shortening the various readings for visual purposes, but there's
no need to do that in passing the data to Homebridge (except to save a
few bytes on the wire).

I also fixed the mimimum value that can be passed for the light level.
The HomeKit spec requires it to be at least 0.0001. See
https://developers.homebridge.io/#/characteristic/CurrentAmbientLightLevel